Add LICENSE file (LGPL v2.1)
This file is included for reference purposes only, and should not be considered
a replacement for the copyright notices contained in individual source files.
I am not the copyright holder of the code distributed in this repository, but
to the best of my knowledge, all the code mirrored here is redistributable
under the LGPL v2.1.